# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/project_templates:
    get:
      tags:
        - ProjectTemplates
      summary: List ProjectTemplates
      description: Get a paginated list of ProjectTemplates with optional filtering and sorting
      operationId: listProjectTemplates
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectTemplateList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ProjectTemplates
      summary: Create ProjectTemplate
      description: Create a new ProjectTemplate
      operationId: createProjectTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectTemplateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/project_template_tasks:
    get:
      tags:
        - ProjectTemplateTasks
      summary: List ProjectTemplateTasks
      description: Get a paginated list of ProjectTemplateTasks with optional filtering and sorting
      operationId: listProjectTemplateTasks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectTemplateTaskList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ProjectTemplateTasks
      summary: Create ProjectTemplateTask
      description: Create a new ProjectTemplateTask
      operationId: createProjectTemplateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectTemplateTaskRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectTemplateTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    ProjectTemplate:
      type: object
      required:
        - id
        - company_id
        - template_name
        - project_type
        - is_active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        template_name:
          type: string
          maxLength: 200
        project_type:
          type: string
        is_active:
          type: boolean
        metadata:
          type: object
...
    CreateProjectTemplateRequest:
      type: object
      required:
        - company_id
        - template_name
        - project_type
        - is_active
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        template_name:
          type: string
          maxLength: 200
        project_type:
          type: string
        is_active:
          type: boolean
        metadata:
          type: object
...
    ProjectTemplateList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectTemplate'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ProjectTemplateTask:
      type: object
      required:
        - id
        - company_id
        - template_id
        - subject
        - expected_time
        - sequence
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        template_id:
          type: string
          format: uuid
        subject:
          type: string
          maxLength: 200
        task_type:
          nullable: true
          type: string
          maxLength: 80
        expected_time:
          type: string
        sequence:
          type: integer
          format: int32
        metadata:
          type: object
...
    CreateProjectTemplateTaskRequest:
      type: object
      required:
        - company_id
        - template_id
        - subject
        - expected_time
        - sequence
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        template_id:
          type: string
          format: uuid
        subject:
          type: string
          maxLength: 200
        task_type:
          nullable: true
          type: string
          maxLength: 80
        expected_time:
          type: string
        sequence:
          type: integer
          format: int32
        metadata:
          type: object
...
    ProjectTemplateTaskList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectTemplateTask'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []