# src/presentation/http/project_template_handler.rs
//! ProjectTemplate REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{ProjectTemplateService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateProjectTemplateDto, UpdateProjectTemplateDto, PatchProjectTemplateDto, ProjectTemplateResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum ProjectTemplateError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for ProjectTemplateError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for ProjectTemplateError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "PROJECTTEMPLATE_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "PROJECTTEMPLATE_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "PROJECTTEMPLATE_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "PROJECTTEMPLATE_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for ProjectTemplate.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /project_templates | List with pagination |
/// | POST | /project_templates | Create new |
/// | GET | /project_templates/:id | Get by ID |
/// | PUT | /project_templates/:id | Full update |
/// | PATCH | /project_templates/:id | Partial update |
/// | DELETE | /project_templates/:id | Soft delete |
/// | POST | /project_templates/bulk | Bulk create |
/// | POST | /project_templates/upsert | Upsert |
/// | GET | /project_templates/trash | List deleted |
/// | POST | /project_templates/:id/restore | Restore |
/// | DELETE | /project_templates/empty | Empty trash |
/// | GET | /project_templates/:id/deleted | Get deleted by ID |
/// | DELETE | /project_templates/trash/:id | Permanent delete from trash |
/// | GET | /project_templates/count | Count active entities |
/// | GET | /project_templates/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(ProjectTemplateService::with_repository(repository));
/// let router = create_project_template_routes(service);
/// ```
pub fn create_project_template_routes(service: Arc<ProjectTemplateService>) -> Router {
    BackboneCrudHandler::<ProjectTemplateService, ProjectTemplate, CreateProjectTemplateDto, UpdateProjectTemplateDto, ProjectTemplateResponseDto>::routes(
        service,
        "/project_templates",
    )
}

/// Create Axum router with only the read (GET) endpoints for ProjectTemplate.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_project_template_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_project_template_read_routes(service: Arc<ProjectTemplateService>) -> Router {
    BackboneCrudHandler::<ProjectTemplateService, ProjectTemplate, CreateProjectTemplateDto, UpdateProjectTemplateDto, ProjectTemplateResponseDto>::read_routes(
        service,
        "/project_templates",
    )
}

/// Create Axum router with only the write (mutation) endpoints for ProjectTemplate.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_project_template_write_routes(service: Arc<ProjectTemplateService>) -> Router {
    BackboneCrudHandler::<ProjectTemplateService, ProjectTemplate, CreateProjectTemplateDto, UpdateProjectTemplateDto, ProjectTemplateResponseDto>::write_routes(
        service,
        "/project_templates",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_project_template_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<ProjectTemplateService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_project_template_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}