# src/presentation/http/project_template_task_handler.rs
//! ProjectTemplateTask REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{ProjectTemplateTaskService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateProjectTemplateTaskDto, UpdateProjectTemplateTaskDto, PatchProjectTemplateTaskDto, ProjectTemplateTaskResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum ProjectTemplateTaskError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for ProjectTemplateTaskError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for ProjectTemplateTaskError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "PROJECTTEMPLATETASK_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "PROJECTTEMPLATETASK_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "PROJECTTEMPLATETASK_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "PROJECTTEMPLATETASK_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for ProjectTemplateTask.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /project_template_tasks | List with pagination |
/// | POST | /project_template_tasks | Create new |
/// | GET | /project_template_tasks/:id | Get by ID |
/// | PUT | /project_template_tasks/:id | Full update |
/// | PATCH | /project_template_tasks/:id | Partial update |
/// | DELETE | /project_template_tasks/:id | Soft delete |
/// | POST | /project_template_tasks/bulk | Bulk create |
/// | POST | /project_template_tasks/upsert | Upsert |
/// | GET | /project_template_tasks/trash | List deleted |
/// | POST | /project_template_tasks/:id/restore | Restore |
/// | DELETE | /project_template_tasks/empty | Empty trash |
/// | GET | /project_template_tasks/:id/deleted | Get deleted by ID |
/// | DELETE | /project_template_tasks/trash/:id | Permanent delete from trash |
/// | GET | /project_template_tasks/count | Count active entities |
/// | GET | /project_template_tasks/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(ProjectTemplateTaskService::with_repository(repository));
/// let router = create_project_template_task_routes(service);
/// ```
pub fn create_project_template_task_routes(service: Arc<ProjectTemplateTaskService>) -> Router {
    BackboneCrudHandler::<ProjectTemplateTaskService, ProjectTemplateTask, CreateProjectTemplateTaskDto, UpdateProjectTemplateTaskDto, ProjectTemplateTaskResponseDto>::routes(
        service,
        "/project_template_tasks",
    )
}

/// Create Axum router with only the read (GET) endpoints for ProjectTemplateTask.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_project_template_task_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_project_template_task_read_routes(service: Arc<ProjectTemplateTaskService>) -> Router {
    BackboneCrudHandler::<ProjectTemplateTaskService, ProjectTemplateTask, CreateProjectTemplateTaskDto, UpdateProjectTemplateTaskDto, ProjectTemplateTaskResponseDto>::read_routes(
        service,
        "/project_template_tasks",
    )
}

/// Create Axum router with only the write (mutation) endpoints for ProjectTemplateTask.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_project_template_task_write_routes(service: Arc<ProjectTemplateTaskService>) -> Router {
    BackboneCrudHandler::<ProjectTemplateTaskService, ProjectTemplateTask, CreateProjectTemplateTaskDto, UpdateProjectTemplateTaskDto, ProjectTemplateTaskResponseDto>::write_routes(
        service,
        "/project_template_tasks",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_project_template_task_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<ProjectTemplateTaskService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_project_template_task_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}